main printing efficiency

+++ AnalysisAndSuch/JetFake/main.cpp
@@ -1044,7 +1044,7 @@ int main(int argc, const char * argv[])
         for (int c=0; c<GangCutCount.size(); c++) cout << GangCutCount[c] << endl;
         cout << "num eejj events b4 cuts: " << numSSdilep << endl;
         cout << deepCuts2[10] << endl << NumEntries << endl;
-        double efficency = deepCuts2[10]/NumEntries;
+        long efficency = deepCuts2[10]/NumEntries;
         cout << "Efficiency: " << efficency << endl;
     }
